sunset sampling support, off-load to s2fft v0.0.1
