+++ s2wav/utils/__init__.py
@@ -1,3 +1,2 @@
 from . import math_functions
-from . import samples
 from . import shapes
